add EcsFwd
