+++ include/ncengine/audio/NcAudio.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "ncengine/audio/AudioSource.h"
-#include "ncengine/ecs/Ecs.h"
+#include "ncengine/ecs/EcsFwd.h"
 #include "ncengine/ecs/Entity.h"
 #include "ncengine/module/Module.h"
 #include "ncengine/type/EngineId.h"
@@ -15,6 +15,8 @@
 
 namespace nc
 {
+class Transform;
+
 namespace config
 {
 struct AudioSettings;
