+++ source/engine/graphics/system/WidgetSystem.cpp
@@ -1,5 +1,11 @@
 #include "WidgetSystem.h"
 #include "ncengine/asset/DefaultAssets.h"
+#include "ncengine/ecs/Ecs.h"
+#include "ncengine/ecs/Transform.h"
+#include "ncengine/graphics/MeshRenderer.h"
+#include "ncengine/graphics/ToonRenderer.h"
+#include "ncengine/graphics/WireframeRenderer.h"
+#include "ncengine/physics/Collider.h"
 #include "asset/AssetService.h"
 
 #include "optick.h"
